Added config.cpp to Loadouts/Cates and split it up into .hpp
