--- Addons/Manticore_Loadout/3CBVehicles/Manticore_Loadout_3CBVehicles.hpp
@@ -1,17 +1,0 @@
-
-//Light Atack
-#include "LandRover.hpp"
-#include "Coyotejackal.hpp"
-//Car
-#include "Panther.hpp"
-//APC
-#include "Buldog.hpp"
-//Heli
-#include "Wildcat.hpp"
-#include "Merlin.hpp"
-#include "Apache.hpp"
-//Naval
-#include "RHIB.hpp"
-
-
-
