+++ Addons/Manticore_Loadout/Crates/Ammo.hpp
@@ -1,0 +1,280 @@
+class Manticore_Loadout_Crate_Ammo_StandardRifle : Box_NATO_Ammo_F
+{
+	displayName = "Ammocrate (G36k(KSK)) (Ammo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+		_standardMag(50, 50);
+		_UGLHE(30);
+		_UGLFlare(20, 10, 20, 0, 0);
+		_UGLSmoke(30, 10, 10, 10, 10, 0, 0)
+
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Kitbag
+		{
+			backpack = "VSM_OGA_OD_Backpack_Kitbag";
+			count = 1;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Ammo_MAAWS_Vanila : Box_NATO_WpsLaunch_F
+{
+	displayName = "Launcher Crate (MAAWS) (Ammo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+		class _xx_MRAWS_HEAT_F
+		{
+			magazine = "MRAWS_HEAT_F";
+			count = 12;
+		};
+		class _xx_MRAWS_HE_F
+		{
+			magazine = "MRAWS_HE_F";
+			count = 6;
+		};
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Kitbag
+		{
+			backpack = "VSM_OGA_OD_Backpack_Kitbag";
+			count = 1;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Ammo_Metis_Vanila : Box_NATO_WpsLaunch_F
+{
+	displayName = "Launcher Crate (Metis-M) (Ammo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+		class _xx_Vorona_HEAT
+		{
+			magazine = "Vorona_HEAT";
+			count = 10;
+		};
+		class _xx_Vorona_HE
+		{
+			magazine = "Vorona_HE";
+			count = 4;
+		};
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Kitbag
+		{
+			backpack = "VSM_OGA_OD_Backpack_Kitbag";
+			count = 1;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Ammo_Stinger_RHS : Box_NATO_WpsLaunch_F
+{
+	displayName = "Launcher Crate (Stinger) (Ammo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+		class _xx_rhs_fim92_mag
+		{
+			magazine = "rhs_fim92_mag";
+			count = 12;
+		};
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Kitbag
+		{
+			backpack = "VSM_OGA_OD_Backpack_Kitbag";
+			count = 1;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Ammo_Mk46_HLC : Box_NATO_Ammo_F
+{
+	displayName = "Ammocrate (Mk46) (Ammo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+		class _xx_hlc_200rnd_556x45_M_SAW
+		{
+			magazine = "hlc_200rnd_556x45_M_SAW";
+			count = 16;
+		};
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+	};
+};
+class Manticore_Loadout_Crate_Ammo_Mk48_HLC : Box_NATO_Ammo_F
+{
+	displayName = "Ammocrate (Mk48) (Ammo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+		class _xx_hlc_100Rnd_762x51_Mdim_M60E4
+		{
+			magazine = "hlc_100Rnd_762x51_Mdim_M60E4";
+			count = 10;
+		};
+		class _xx_hlc_100Rnd_762x51_M_M60E4
+		{
+			magazine = "hlc_100Rnd_762x51_M_M60E4";
+			count = 10;
+		};
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+	};
+};
+class Manticore_Loadout_Crate_Ammo_M14_HLC : Box_NATO_Ammo_F
+{
+	displayName = "Ammocrate (Troy M14) (Ammo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+		class _xx_hlc_20Rnd_762x51_T_M14
+		{
+			magazine = "hlc_20Rnd_762x51_T_M14";
+			count = 20;
+		};
+		class _xx_hlc_20Rnd_762x51_B_M14
+		{
+			magazine = "hlc_20Rnd_762x51_B_M14";
+			count = 20;
+		};
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+	};
+};
+class Manticore_Loadout_Crate_Ammo_Mk11_RHS : Box_NATO_Ammo_F
+{
+	displayName = "Ammocrate (Mk 11 Mod 0) (Ammo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+		class _xx_rhsusf_20Rnd_762x51_SR25_m993_Mag
+		{
+			magazine = "rhsusf_20Rnd_762x51_SR25_m993_Mag";
+			count = 10;
+		};
+		class _xx_rhsusf_20Rnd_762x51_SR25_m62_Mag
+		{
+			magazine = "rhsusf_20Rnd_762x51_SR25_m62_Mag";
+			count = 10;
+		};
+		class _xx_rhsusf_20Rnd_762x51_SR25_m118_special_Mag
+		{
+			magazine = "rhsusf_20Rnd_762x51_SR25_m118_special_Mag";
+			count = 10;
+		};
+		class _xx_ACE_10Rnd_762x51_M118LR_Mag
+		{
+			magazine = "ACE_10Rnd_762x51_M118LR_Mag";
+			count = 10;
+		};
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+	};
+};
+class Manticore_Loadout_Crate_Ammo_M27_RHS : Box_NATO_Ammo_F
+{
+	displayName = "Ammocrate (M27 IAR) (Ammo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+		class _xx_rhs_mag_100Rnd_556x45_M855A1_cmag_mixed
+		{
+			magazine = "rhs_mag_100Rnd_556x45_M855A1_cmag_mixed";
+			count = 20;
+		};
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+	};
+};
