+++ Addons/Manticore_Loadout/Crates/Demo.hpp
@@ -1,0 +1,191 @@
+class Manticore_Loadout_Crate_Demo_Demo : Box_NATO_Support_F
+{
+	displayName = "Support Crate (Demo) (Demo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+
+	};
+	class TransportMagazines
+	{
+
+	};
+	class TransportItems
+	{
+		class _xx_DemoCharge_Remote_Mag
+		{
+			name = "DemoCharge_Remote_Mag";
+			count = 10;
+		};
+		class _xx_SatchelCharge_Remote_Mag
+		{
+			name = "SatchelCharge_Remote_Mag";
+			count = 5;
+		};
+		class _xx_ACE_Clacker
+		{
+			name = "ACE_Clacker";
+			count = 1;
+		};
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Kitbag
+		{
+			backpack = "VSM_OGA_OD_Backpack_Kitbag";
+			count = 1;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Demo_Mine : Box_NATO_Support_F
+{
+	displayName = "Support Crate (Mine) (Demo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+
+	};
+	class TransportMagazines
+	{
+
+	};
+	class TransportItems
+	{
+		class _xx_SLAMDirectionalMine_Wire_Mag
+		{
+			name = "SLAMDirectionalMine_Wire_Mag";
+			count = 5;
+		};
+		class _xx_ATMine_Range_Mag
+		{
+			name = "ATMine_Range_Mag";
+			count = 7;
+		};
+		class _xx_APERSMineDispenser_Mag
+		{
+			name = "APERSBoundingMine_Range_Mag";
+			count = 3;
+		};
+		class _xx_ClaymoreDirectionalMine_Remote_Mag
+		{
+			name = "ClaymoreDirectionalMine_Remote_Mag";
+			count = 5;
+		};
+		class _xx_rhs_mine_msk40p_blue_mag
+		{
+			name = "rhs_mine_msk40p_blue_mag";
+			count = 5;
+		};
+		class _xx_rhs_mine_msk40p_red_mag
+		{
+			name = "rhs_mine_msk40p_red_mag";
+			count = 5;
+		};
+		class _xx_APERSMine_Range_Mag
+		{
+			name = "APERSMine_Range_Mag";
+			count = 10;
+		};
+		class _xx_APERSBoundingMine_Range_Mag
+		{
+			name = "APERSBoundingMine_Range_Mag";
+			count = 4;
+		};
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Kitbag
+		{
+			backpack = "VSM_OGA_OD_Backpack_Kitbag";
+			count = 1;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Demo_DemoMine : Box_NATO_Support_F
+{
+	displayName = "Support Crate (Demo and mine) (Demo) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+
+	};
+	class TransportMagazines
+	{
+
+	};
+	class TransportItems
+	{
+		class _xx_DemoCharge_Remote_Mag
+		{
+			name = "DemoCharge_Remote_Mag";
+			count = 5;
+		};
+		class _xx_SatchelCharge_Remote_Mag
+		{
+			name = "SatchelCharge_Remote_Mag";
+			count = 2;
+		};
+		class _xx_ACE_Clacker
+		{
+			name = "ACE_Clacker";
+			count = 1;
+		};
+		class _xx_SLAMDirectionalMine_Wire_Mag
+		{
+			name = "SLAMDirectionalMine_Wire_Mag";
+			count = 5;
+		};
+		class _xx_ATMine_Range_Mag
+		{
+			name = "ATMine_Range_Mag";
+			count = 4;
+		};
+		class _xx_APERSMineDispenser_Mag
+		{
+			name = "APERSBoundingMine_Range_Mag";
+			count = 2;
+		};
+		class _xx_APERSMine_Range_Mag
+		{
+			name = "APERSMine_Range_Mag";
+			count = 10;
+		};
+		class _xx_APERSBoundingMine_Range_Mag
+		{
+			name = "APERSBoundingMine_Range_Mag";
+			count = 5;
+		};
+		class _xx_rhs_mine_msk40p_blue_mag
+		{
+			name = "rhs_mine_msk40p_blue_mag";
+			count = 5;
+		};
+		class _xx_rhs_mine_msk40p_red_mag
+		{
+			name = "rhs_mine_msk40p_red_mag";
+			count = 5;
+		};
+		class _xx_ClaymoreDirectionalMine_Remote_Mag
+		{
+			name = "ClaymoreDirectionalMine_Remote_Mag";
+			count = 5;
+		};
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Kitbag
+		{
+			backpack = "VSM_OGA_OD_Backpack_Kitbag";
+			count = 1;
+		};
+	};
+};
