+++ Addons/Manticore_Loadout/Crates/GrenadesSmokFlare.hpp
@@ -1,0 +1,204 @@
+class Manticore_Loadout_Crate_Grenade_Grenade : Box_NATO_Support_F
+{
+	displayName = "Support Crate (Grenade) (Grenade) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+
+	};
+	class TransportMagazines
+	{
+
+	};
+	class TransportItems
+	{
+		_GrenadeOffensive(40, 20, 10);
+		_GrenadeIR(20);
+		_GrenadeSmoke(10, 20, 10, 10, 5, 5, 10);
+		_GrenadeFlare(10, 0, 10, 10);
+
+
+		class _xx_ACE_Chemlight_HiBlue
+		{
+			name = "ACE_Chemlight_HiBlue";
+			count = 10;
+		};
+		class _xx_ACE_Chemlight_HiGreen
+		{
+			name = "ACE_Chemlight_HiGreen";
+			count = 10;
+		};
+		class _xx_ACE_Chemlight_UltraHiOrange
+		{
+			name = "ACE_Chemlight_UltraHiOrange";
+			count = 10;
+		};
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Compact
+		{
+			backpack = "VSM_OGA_OD_Backpack_Compact";
+			count = 1;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Grenade_FlareChem : Box_NATO_Support_F
+{
+	displayName = "Support Crate (Flare and Chem) (Grenade) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+
+	};
+	class TransportMagazines
+	{
+
+	};
+	class TransportItems
+	{
+		_GrenadeFlare(30, 30, 30, 30);
+
+		class _xx_ACE_Chemlight_IR
+		{
+			name = "ACE_Chemlight_IR";
+			count = 30;
+		};
+		class _xx_Chemlight_blue
+		{
+			name = "Chemlight_blue";
+			count = 30;
+		};
+		class _xx_Chemlight_green
+		{
+			name = "Chemlight_green";
+			count = 30;
+		};
+		class _xx_ACE_Chemlight_HiRed
+		{
+			name = "ACE_Chemlight_HiRed";
+			count = 30;
+		};
+		class _xx_ACE_Chemlight_HiBlue
+		{
+			name = "ACE_Chemlight_HiBlue";
+			count = 30;
+		};
+		class _xx_ACE_Chemlight_HiGreen
+		{
+			name = "ACE_Chemlight_HiGreen";
+			count = 30;
+		};
+		class _xx_ACE_Chemlight_UltraHiOrange
+		{
+			name = "ACE_Chemlight_UltraHiOrange";
+			count = 30;
+		};
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Compact
+		{
+			backpack = "VSM_OGA_OD_Backpack_Compact";
+			count = 1;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Grenade_SmokFlareChem : Box_NATO_Support_F
+{
+	displayName = "Support Crate (Smok, Flare and Chem) (Grenade) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+
+	};
+	class TransportMagazines
+	{
+
+	};
+	class TransportItems
+	{
+		_GrenadeSmoke(10, 30, 10, 10, 10, 10, 10);
+		_GrenadeFlare(10, 10, 10, 10);
+		class _xx_ACE_Chemlight_IR
+		{
+			name = "ACE_Chemlight_IR";
+			count = 20;
+		};
+		class _xx_Chemlight_blue
+		{
+			name = "Chemlight_blue";
+			count = 20;
+		};
+		class _xx_Chemlight_green
+		{
+			name = "Chemlight_green";
+			count = 20;
+		};
+		class _xx_ACE_Chemlight_HiRed
+		{
+			name = "ACE_Chemlight_HiRed";
+			count = 20;
+		};
+		class _xx_ACE_Chemlight_HiBlue
+		{
+			name = "ACE_Chemlight_HiBlue";
+			count = 20;
+		};
+		class _xx_ACE_Chemlight_HiGreen
+		{
+			name = "ACE_Chemlight_HiGreen";
+			count = 20;
+		};
+		class _xx_ACE_Chemlight_UltraHiOrange
+		{
+			name = "ACE_Chemlight_UltraHiOrange";
+			count = 20;
+		};
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Compact
+		{
+			backpack = "VSM_OGA_OD_Backpack_Compact";
+			count = 1;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Grenade_Smok : Box_NATO_Support_F
+{
+	displayName = "Support Crate (Smok) (Grenade) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+
+	};
+	class TransportMagazines
+	{
+
+	};
+	class TransportItems
+	{
+		_GrenadeSmoke(20, 60, 20, 20, 20, 20, 20);
+	};
+	class TransportBackpacks
+	{
+		class _xx_VSM_OGA_OD_Backpack_Compact
+		{
+			backpack = "VSM_OGA_OD_Backpack_Compact";
+			count = 1;
+		};
+	};
+};
