+++ Addons/Manticore_Loadout/Crates/Static.hpp
@@ -1,0 +1,123 @@
+class Manticore_Loadout_Crate_Static_M2_RHS : Box_NATO_Support_F
+{
+	displayName = "Support Crate (M2) (Static) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+		class _xx_RHS_M2_Gun_Bag
+		{
+			backpack = "RHS_M2_Gun_Bag";
+			count = 2;
+		};
+		class _xx_RHS_M2_Tripod_Bag
+		{
+			backpack = "RHS_M2_Tripod_Bag";
+			count = 2;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Static_Mk19_RHS : Box_NATO_Support_F
+{
+	displayName = "Support Crate (Mk19) (Static) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+		class _xx_RHS_Mk19_Gun_Bag
+		{
+			backpack = "RHS_Mk19_Gun_Bag";
+			count = 2;
+		};
+		class _xx_RHS_Mk19_Tripod_Bag
+		{
+			backpack = "RHS_Mk19_Tripod_Bag";
+			count = 2;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Static_TOW_RHS : Box_NATO_Support_F
+{
+	displayName = "Support Crate (TOW) (Static) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+		class _xx_rhs_Tow_Gun_Bag
+		{
+			backpack = "rhs_Tow_Gun_Bag";
+			count = 2;
+		};
+		class _xx_rhs_TOW_Tripod_Bag
+		{
+			backpack = "rhs_TOW_Tripod_Bag";
+			count = 2;
+		};
+	};
+};
+class Manticore_Loadout_Crate_Static_Mk6_Vanila : Box_NATO_Support_F
+{
+	displayName = "Support Crate (Mk6 Mortar) (Static) (TFM)";
+	scope = 2;
+	editorCategory = "Manticore_Categories";
+	editorSubcategory = "Manticore_Categories_Supply";
+
+	class TransportWeapons
+	{
+	};
+	class TransportMagazines
+	{
+
+	};
+	class TransportItems
+	{
+	};
+	class TransportBackpacks
+	{
+		class _xx_B_Mortar_01_support_F
+		{
+			backpack = "B_Mortar_01_support_F";
+			count = 2;
+		};
+		class _xx_B_Mortar_01_weapon_F
+		{
+			backpack = "B_Mortar_01_weapon_F";
+			count = 2;
+		};
+	};
+};
