--- Addons/Manticore_Loadout/RHS/Manticore_Loadout_RHS.hpp
@@ -1,8 +1,0 @@
-
-
-#include "Humvee.hpp"
-#include "MRAP.hpp"
-#include "APC.hpp"
-#include "IFV.hpp"
-#include "Heli.hpp"
-#include "Naval.hpp"
