--- Addons/Manticore_Loadout/Vanila/Manticore_Loadout_Vanila.hpp
@@ -1,24 +1,0 @@
-
-//Light Attack
-	#include "Polaris.hpp"
-	#include "Quadbike.hpp"
-//Heli
-	#include "Wildcat.hpp"
-	//#include "Merlin.hpp"
-	#include "LittleBird.hpp"
-//APC IFV
-	#include "Badger.hpp"
-	#include "Namer.hpp"
-	#include "Nemmera.hpp"
-//AA
-	#include "Bardelas.hpp"
-//Tank
-	#include "Wiesel2.hpp"
-	#include "Leopard.hpp"
-	#include "Rooikat.hpp"
-//Atrillery
-	#include "Artillery.hpp"
-//Naval
-	#include "Speedboat.hpp"
-//Drone
-	#include "Drone.hpp"
